// src/models/editedMeasurement.model.ts
export interface EditedMeasurement {
    fragment?: string;
    msgId?: string;
    series?: string;
    unit?: string;
    value?: number;
    index?: number;
    msmt?: {
        fragment?: string;
        msgId?: string;
        series?: string;
        unit?: string;
        value?: number;
    };
}